requirements.txt
streamlit==1.36.0
openai==1.35.15
python-dotenv==1.0.1
requests==2.32.3